==> web_tool/multi-stop_route_optimization/code/python/05_Output.py <==
import qrcode
from fpdf import FPDF
import os
import pandas as pd
import time

f = open("output/04/final.txt","r")
address_list = []
for i,txt in enumerate(f):
    address = txt.replace("\n","")
    address_list = address_list + [address]
    img = qrcode.make(address)
    type(img)  # qrcode.image.pil.PilImage
    img.save("output/05/_images/"+str(i)+".png")
f.close()


# =============================================================================
# Add table
# =============================================================================
# Import FPDF class
pdf=FPDF(format='letter', unit='in')

# Effective page width, or just epw
epw = pdf.w - 2*pdf.l_margin
qr_code_size = epw/3

start = time.time()
for ii,address in enumerate(address_list):
    end = time.time()
    elapsed_time = end - start
    if elapsed_time>1000:
        raise("taking too long 4")
        
    pdf.add_page()
     
    
    # =============================================================================
    # Title
    # =============================================================================
    pdf.set_text_color(0,0,0)
    pdf.set_font('Times','B',14.0)
    pdf.cell(epw, 0.0, 'Route '+str(ii+1)+' List', align='C')
    pdf.set_font('Times','',10.0)
    pdf.ln(0.5)
     
    
    # =============================================================================
    # 
    # =============================================================================
    df = pd.read_csv("output/04/"+str(ii)+".csv")
    th = pdf.font_size + 0.1
    col_width_pct = [0.05,0.65,0.15,0.15]
    
    pdf.set_font('Times','B',10.0)
    pdf.set_text_color(0,0,0)
    
    for j,t in enumerate(["#","Input Address/Coordinate","Lat","Lon"]):
        pdf.cell(epw*col_width_pct[j], th, str(t), border=1,align="C")
    pdf.ln(th)
    
    pdf.set_font('Times','',10.0)
    pdf.set_text_color(0,0,0)
    for i in range(len(df)):
        for j,t in enumerate(df.iloc[i]):
            pdf.cell(epw*col_width_pct[j], th, str(t)[:50], border=1,align="C")
        pdf.ln(th)
    
    
    
    # pdf.link(x=col_width1, w=col_width2, y=th, h=th, link="https://github.com/PyFPDF/fpdf2")
    
    pdf.set_xy(pdf.l_margin, pdf.h - qr_code_size - pdf.b_margin*2)
    pdf.set_font('Times','U',14.0)
    pdf.set_text_color(153,0,153)
    pdf.set_fill_color(255, 255, 255)
    pdf.cell(epw, th, "Route "+str(ii+1)+" Link", border=0, fill=True, link=address,align="C")
    
    # =============================================================================
    # Add QR code
    # =============================================================================
    
    pdf.set_xy(pdf.w/2-qr_code_size/2, pdf.h - qr_code_size - pdf.b_margin*1.5)
    pltx=("output/05/_images/"+str(ii)+".png")
    pdf.image(pltx,  link='', type='', w=qr_code_size)
    
    pdf.set_xy(pdf.l_margin, pdf.h - pdf.b_margin*1.5)
    pdf.set_font('Times','U',10.0)
    pdf.set_text_color(0,0,0)
    pdf.cell(epw, th, "Generated by Route.BongBong.Store", border=0, fill=False, link="https://Route.BongBong.Store/",align="C")


# =============================================================================
# output
# =============================================================================
pdf.output('output/05/suggested_route.pdf','F')
